src/requirements.txt
pytz~=2020.4
tzlocal~=2.1
APScheduler~=3.6.3
togglCli